changed to_i to to_f for duration multiplier

+++ app/models/availability.rb
@@ -16,11 +16,11 @@ class Availability < ApplicationRecord
   validates :duration, presence: true, length: { maximum: 2 }, format: { with: VALID_DURATION_REGEX }, inclusion: {in: VALID_DURATIONS}
   
   def end_before_close
-     if (!(time.to_s=~VALID_TIME_REGEX).nil?)
-   if ((time.to_s.to_time)+(3600*duration.to_i)>("18:00".to_time))
-     errors.add(:duration, "cant end after closing")
-   end
-   end
+    if (!(time.to_s=~VALID_TIME_REGEX).nil?)
+      if ((time.to_s.to_time)+(3600*duration.to_f)>("18:00".to_time))
+        errors.add(:duration, "cant end after closing")
+      end
+    end
   end
 
 end
