+++ app/models/free_play_room.rb
@@ -20,7 +20,7 @@ class FreePlayRoom < ApplicationRecord
   
   def end_before_close
     if (!(time.to_s=~VALID_TIME_REGEX).nil?)
-      if ((time.to_s.to_time)+(3600*duration.to_i)>("18:00".to_time))
+      if ((time.to_s.to_time)+(3600*duration.to_f)>("18:00".to_time))
         errors.add(:duration, "cant end after closing")
       end
     end
